+++ R/03-modelResults3DKernel.R
@@ -164,6 +164,8 @@ modelResults3DKernelUI <- function(id, title = ""){
         conditionalPanel(
           condition = conditionPlot(ns("DistMap")),
           textOutput(ns("centerEstimate"), container = function(...) div(..., style = "text-align:center;")),
+          tags$br(),
+          tags$br(),
           div(plotExportButton(ns("export"))),
           conditionalPanel(
             condition = "input.mapType == 'Map'",
@@ -248,25 +250,32 @@ modelResults3DKernelUI <- function(id, title = ""){
                        selected = "Map"),
           conditionalPanel(
             condition = "input.mapType == 'Time course'",
+            ns = ns,
+            tags$hr(),
             selectInput(inputId = ns("intervalType"), label = "Uncertainty Interval Type",
                         choices = list("none" = "1",
                                        "1 SE" = "2",
                                        "2 SE" = "6"), selected = "2"),
-          checkboxInput(inputId = ns("pointsTime"),
-                        label = "Show nearby points",
-                        value = TRUE, width = "100%"),
-          conditionalPanel(
-            condition = "input.pointsTime == true",
-            checkboxInput(inputId = ns("intTime"),
-                          label = "Show nearby points unc. intervals",
-                          value = FALSE, width = "100%"),
-            sliderInput(inputId = ns("rangePointsTime"),
-                        label = "Show nearby points / intervals range in km",
-                        min = 10, max = 2500, value = 250, step = 10),
-            ns = ns),
-          ns = ns),
+            checkboxInput(inputId = ns("pointsTime"),
+                          label = "Show nearby points",
+                          value = TRUE, width = "100%"),
+            conditionalPanel(
+              condition = "input.pointsTime == true",
+              ns = ns,
+              checkboxInput(inputId = ns("intTime"),
+                            label = "Show nearby points unc. intervals",
+                            value = FALSE, width = "100%"),
+              sliderInput(inputId = ns("rangePointsTime"),
+                          label = "Show nearby points / intervals range in km",
+                          min = 10, max = 2500, value = 250, step = 10)
+              ),
+            formatTimeCourseUI(ns("timeCourseFormat")),
+            tags$hr()
+            ),
           conditionalPanel(
             condition = "input.mapType != 'Time course'",
+            ns = ns,
+            tags$hr(),
             radioButtons(inputId = ns("terrestrial"), label = "", inline = TRUE,
                          choices = list("Terrestrial " = 1, "All" = 3, "Aquatic" = -1),
                          selected = 1),
@@ -414,7 +423,13 @@ modelResults3DKernelUI <- function(id, title = ""){
             sliderInput(inputId = ns("ncol"),
                         label = "Approximate number of colour levels",
                         min = 4, max = 50, value = 50, step = 2, width = "100%"),
-            ns = ns),
+            conditionalPanel(
+              condition = "input.mapType == 'Map'",
+              ns = ns,
+              centerEstimateUI(ns("centerEstimateParams"))
+            ),
+            tags$hr()
+            ),
           checkboxInput(inputId = ns("smoothCols"),
                         label = "Smooth color transition",
                         value = FALSE, width = "100%"),
@@ -448,24 +463,13 @@ modelResults3DKernelUI <- function(id, title = ""){
             colourInput(inputId = ns("fontCol"),
                         label = "Colour of font",
                         value = "#2C2161"), ns = ns),
-          numericInput(inputId = ns("centerY"),
-                       label = "Center point latitude",
-                       min = -180, max = 180, value = c(), step = 0.5, width = "100%"),
-          numericInput(inputId = ns("centerX"),
-                       label = "Center point longitude",
-                       min = -90, max = 90, value = c(), step = 0.5, width = "100%"),
-          conditionalPanel(
-            condition = "input.timeCourse != 'mapType'",
-            sliderInput(inputId = ns("Radius"),
-                        label = "Radius (km)",
-                        min = 10, max = 300, value = 100, step = 10, width = "100%"),
-            ns = ns),
           sliderInput(inputId = ns("AxisSize"),
                       label = "Axis title font size",
                       min = 0.1, max = 3, value = 1, step = 0.1, width = "100%"),
           sliderInput(inputId = ns("AxisLSize"),
                       label = "Axis label font size",
                       min = 0.1, max = 3, value = 1, step = 0.1, width = "100%"),
+
           batchPointEstimatesUI(ns("batch"))
         )
     )
@@ -516,19 +520,7 @@ modelResults3DKernel <- function(input, output, session, isoData, savedMaps, fru
 
 
   output$centerEstimate <- renderText({
-    if (is.na(input$centerY) | is.na(input$centerX) | is.na(input$Radius) | input$mapType != "Map") return("")
-
-    if (is.na(values$meanCenter) | is.na(values$sdCenter)) {
-      return("Cannot compute mean and sd at your provided coordinates.
-             Please raise the plot resolution or radius such that estimates within the radius are available.")
-    }
-
-    paste0("Mean: ", values$meanCenter,
-           ", Standard error of the mean: ", values$sdCenter,
-           "  at coordinates ",  "(",
-           input$centerY, "\u00B0, " , input$centerX,
-           "\u00B0) for a ", round(input$Radius, 3),
-           " km radius")
+    centerEstimate$text()
   })
 
 
@@ -879,6 +871,12 @@ modelResults3DKernel <- function(input, output, session, isoData, savedMaps, fru
     return(pointDat2D())
   })
 
+  centerEstimate <- centerEstimateServer("centerEstimateParams",
+                                         meanCenter = reactive(values$meanCenter),
+                                         sdCenter = reactive(values$sdCenter),
+                                         mapType = reactive(input$mapType))
+
+  formatTimeCourse <- formatTimeCourseServer("timeCourseFormat")
 
   userInputTime <- sliderAndNumericInputServer("timeExtended",
                                                value = reactive(dateExtent$mean),
@@ -982,9 +980,9 @@ modelResults3DKernel <- function(input, output, session, isoData, savedMaps, fru
                        trange = input$trange,
                        independent = isolate(Independent()),
                        resolution = input$resolution,
-                       centerX = input$centerX,
-                       centerY = input$centerY,
-                       Radius = input$Radius,
+                       centerX = centerEstimate$centerX(),
+                       centerY = centerEstimate$centerY(),
+                       Radius = centerEstimate$radius(),
                        rangey = c(input$rangezMin, input$rangezMax),
                        pointDat = pointDat,
                        seType = input$intervalType,
@@ -993,6 +991,7 @@ modelResults3DKernel <- function(input, output, session, isoData, savedMaps, fru
                        rangePointsTime = input$rangePointsTime,
                        intTime = input$intTime,
                        limitz = NULL,
+                       formatTimeCourse = formatTimeCourse(),
                        ...)
       } else {
       if(input$mapType == "Time intervals by cluster"){
@@ -1036,9 +1035,9 @@ modelResults3DKernel <- function(input, output, session, isoData, savedMaps, fru
           fontSize = input$fontSize,
           fontType = input$fontType,
           fontCol = input$fontCol,
-          centerX = input$centerX,
-          centerY = input$centerY,
-          Radius = input$Radius,
+          centerX = centerEstimate$centerX(),
+          centerY = centerEstimate$centerY(),
+          Radius = centerEstimate$radius(),
           terrestrial = input$terrestrial,
           colors = input$Colours,
           reverseColors = input$reverseCols,